# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type EmailCategory = (typeof EmailCategory)[keyof typeof EmailCategory];

export const EmailCategory = {
  PRIMARY: "PRIMARY",
  CRITICAL: "CRITICAL",
  TRANSACTIONS: "TRANSACTIONS",
  PROMOTIONS: "PROMOTIONS",
  SOCIAL: "SOCIAL",
  LOW_PRIORITY: "LOW_PRIORITY",
} as const;

export type EmailUrgency = (typeof EmailUrgency)[keyof typeof EmailUrgency];

export const EmailUrgency = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export interface Email {
  id: string;
  gmailId?: string;
  threadId?: string;
  subject: string;
  from: string;
  fromEmail: string;
  to?: string;
  snippet: string;
  body?: string;
  isRead: boolean;
  isStarred: boolean;
  labels: string[];
  receivedAt: string;
  category: EmailCategory;
  priorityScore: number;
  urgency: EmailUrgency;
  bundledCount?: number;
}

export type AiDecisionCategory =
  (typeof AiDecisionCategory)[keyof typeof AiDecisionCategory];

export const AiDecisionCategory = {
  PRIMARY: "PRIMARY",
  CRITICAL: "CRITICAL",
  TRANSACTIONS: "TRANSACTIONS",
  PROMOTIONS: "PROMOTIONS",
  SOCIAL: "SOCIAL",
  LOW_PRIORITY: "LOW_PRIORITY",
} as const;

export type AiDecisionUrgency =
  (typeof AiDecisionUrgency)[keyof typeof AiDecisionUrgency];

export const AiDecisionUrgency = {
  critical: "critical",
  high: "high",
  medium: "medium",
  low: "low",
} as const;

export type AiDecisionRecommendedAction =
  (typeof AiDecisionRecommendedAction)[keyof typeof AiDecisionRecommendedAction];

export const AiDecisionRecommendedAction = {
  reply: "reply",
  ignore: "ignore",
  archive: "archive",
  track: "track",
  read_later: "read_later",
} as const;

export interface AiDecision {
  id: string;
  emailId: string;
  category: AiDecisionCategory;
  priorityScore: number;
  urgency: AiDecisionUrgency;
  recommendedAction: AiDecisionRecommendedAction;
  confidence: number;
  reason: string;
  summary?: string;
  keyPoints?: string[];
  createdAt?: string;
}

export interface EmailWithDecision {
  email: Email;
  decision?: AiDecision;
}

export interface EmailListResponse {
  emails: EmailWithDecision[];
  total: number;
  hasMore: boolean;
}

export type InboxSummaryCategoryCounts = { [key: string]: number };

export interface InboxSummary {
  needsActionCount: number;
  paymentsCount: number;
  criticalCount: number;
  unreadCount: number;
  categoryCounts: InboxSummaryCategoryCounts;
}

export interface CreateDecisionRequest {
  emailId: string;
  forceRefresh?: boolean;
}

export type LogActionRequestAction =
  (typeof LogActionRequestAction)[keyof typeof LogActionRequestAction];

export const LogActionRequestAction = {
  open: "open",
  reply: "reply",
  ignore: "ignore",
  archive: "archive",
  star: "star",
  mark_read: "mark_read",
  override_decision: "override_decision",
} as const;

export type LogActionRequestDecisionOverride =
  (typeof LogActionRequestDecisionOverride)[keyof typeof LogActionRequestDecisionOverride];

export const LogActionRequestDecisionOverride = {
  reply: "reply",
  ignore: "ignore",
  archive: "archive",
  track: "track",
  read_later: "read_later",
} as const;

export interface LogActionRequest {
  emailId: string;
  action: LogActionRequestAction;
  decisionOverride?: LogActionRequestDecisionOverride;
  timeSpentMs?: number;
}

export interface ActionResponse {
  success: boolean;
  memoryUpdated: boolean;
}

export type SyncStatusResponseStatus =
  (typeof SyncStatusResponseStatus)[keyof typeof SyncStatusResponseStatus];

export const SyncStatusResponseStatus = {
  idle: "idle",
  syncing: "syncing",
  error: "error",
} as const;

export interface SyncStatusResponse {
  status: SyncStatusResponseStatus;
  lastSyncAt?: string;
  emailsSynced: number;
  message?: string;
}

export type GetEmailsParams = {
  category?: GetEmailsCategory;
  limit?: number;
  offset?: number;
};

export type GetEmailsCategory =
  (typeof GetEmailsCategory)[keyof typeof GetEmailsCategory];

export const GetEmailsCategory = {
  PRIMARY: "PRIMARY",
  CRITICAL: "CRITICAL",
  TRANSACTIONS: "TRANSACTIONS",
  PROMOTIONS: "PROMOTIONS",
  SOCIAL: "SOCIAL",
  LOW_PRIORITY: "LOW_PRIORITY",
} as const;