# lib/api-zod/src/generated/types/aiDecision.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { AiDecisionCategory } from "./aiDecisionCategory";
import type { AiDecisionRecommendedAction } from "./aiDecisionRecommendedAction";
import type { AiDecisionUrgency } from "./aiDecisionUrgency";

export interface AiDecision {
  id: string;
  emailId: string;
  category: AiDecisionCategory;
  priorityScore: number;
  urgency: AiDecisionUrgency;
  recommendedAction: AiDecisionRecommendedAction;
  confidence: number;
  reason: string;
  summary?: string;
  keyPoints?: string[];
  createdAt?: string;
}