# lib/api-zod/src/generated/types/email.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailCategory } from "./emailCategory";
import type { EmailUrgency } from "./emailUrgency";

export interface Email {
  id: string;
  gmailId?: string;
  threadId?: string;
  subject: string;
  from: string;
  fromEmail: string;
  to?: string;
  snippet: string;
  body?: string;
  isRead: boolean;
  isStarred: boolean;
  labels: string[];
  receivedAt: string;
  category: EmailCategory;
  priorityScore: number;
  urgency: EmailUrgency;
  bundledCount?: number;
}