# lib/api-zod/src/generated/types/emailListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { EmailWithDecision } from "./emailWithDecision";

export interface EmailListResponse {
  emails: EmailWithDecision[];
  total: number;
  hasMore: boolean;
}