# lib/api-zod/src/generated/types/emailWithDecision.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { AiDecision } from "./aiDecision";
import type { Email } from "./email";

export interface EmailWithDecision {
  email: Email;
  decision?: AiDecision;
}