# lib/api-zod/src/generated/types/inboxSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { InboxSummaryCategoryCounts } from "./inboxSummaryCategoryCounts";

export interface InboxSummary {
  needsActionCount: number;
  paymentsCount: number;
  criticalCount: number;
  unreadCount: number;
  categoryCounts: InboxSummaryCategoryCounts;
}