# lib/api-zod/src/generated/types/logActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI Email Copilot API
 * OpenAPI spec version: 0.1.0
 */
import type { LogActionRequestAction } from "./logActionRequestAction";
import type { LogActionRequestDecisionOverride } from "./logActionRequestDecisionOverride";

export interface LogActionRequest {
  emailId: string;
  action: LogActionRequestAction;
  decisionOverride?: LogActionRequestDecisionOverride;
  timeSpentMs?: number;
}